% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_output.R
\name{cwt_output}
\alias{cwt_output}
\title{Common Weight Output}
\usage{
cwt_output(
  data,
  outcomevars,
  outcomenames = NULL,
  geovar,
  yearvar,
  weightvar,
  years = "all",
  weightopt,
  weightyear = NULL,
  age = NULL,
  agevar = NA,
  ci = FALSE,
  cilevel = 0.95,
  clustvar = NULL,
  stratvar = NULL,
  countryname,
  statelabels = NULL,
  palette = NULL
)
}
\arguments{
\item{data}{Dataset to analyze. Should have one row per respondent and contain survey data from at least two years.}

\item{outcomevars}{Names of dichotomous outcome variable(s) coded 0/1/NA. Provide as a single variable name e.g. "dpt3_vx" or as multiple variables in a concatenated list, e.g. c("dpt3_vx", "mcv1_vx")}

\item{outcomenames}{Names of the outcomes in \code{outcomevars} to use in tables and plots; must be the same length as \code{outcomevars}. Optional, defaults to NULL. E.g. c("DPT3", "MCV1")}

\item{geovar}{Name of geography variable in \code{data}. Must have the same levels across all years in \code{years}.}

\item{yearvar}{Variable in dataset indicating survey year}

\item{weightvar}{Variable in dataset containing survey weights for each respondent}

\item{years}{Data years to analyze. Defaults to "all", which will analyze all years in \code{yearvar}. To analyze a subset of years, provide a vector of years to include, e.g. c(2000, 2005, 2010)}

\item{weightopt}{Which year(s) to use as common weights. "earliest" will calculate outcomes post-stratified to the earliest year in \code{years}. "previous" will calculate outcomes for each survey using the weights from the previous survey year. "custom" allows the user to select a year to use weights from.}

\item{weightyear}{Year to use for post-stratifying if \code{weightopt = "custom"}}

\item{age}{Age to filter by. Defaults to NULL. Provide as a single value of \code{agevar} to keep or provide as a pair of values for minimum and maximum ages, e.g. \code{age = c(12, 23)} will keep \code{agevar >= 12} and \code{<= 23}.}

\item{agevar}{Age variable in dataset; used if \code{age} is not NULL. Defaults to NA.}

\item{ci}{a logical value indicating whether confidence intervals should be calculated for outcomes.  Defaults to FALSE}

\item{cilevel}{Defaults to 0.95}

\item{clustvar}{Variable in the dataset identifying which cluster each respondent belongs to. Used in confidence interval calculations to take survey design into account. Provide if \code{ci = TRUE}. Defaults to NULL.}

\item{stratvar}{Variable in the dataset identifying which stratum each respondent belongs to. Used in confidence interval calculations to take survey design into account. Provide if \code{ci = TRUE}. Defaults to NULL.}

\item{countryname}{Character, name of the country (or other top-level grouping) that \code{data} is from}

\item{statelabels}{Character vector of labels corresponding to the levels of \code{geovar}. Defaults to NULL}

\item{palette}{Color palette for the levels of \code{geovar} in output plots. Defaults to NULL, in which case a default ggplot2 palette will be used.}
}
\description{
Analyze multiple survey datasets: calculate changes in state weights over time, post-stratify to specified year(s), produce tables and plots
}
\examples{
cwt_output()
}
\keyword{survey}
