% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fauxdata}
\alias{fauxdata}
\title{Harmonia survey data}
\format{
A data frame with 9768 rows and 9 variables:
\describe{
\item{geo}{districts of Harmonia}
\item{year}{survey year}
\item{stratumid}{numeric district ID}
\item{clusterid}{survey cluster ID}
\item{age}{respondent age in years}
\item{psweight}{survey weight for each respondent}
\item{has_card}{binary outcome: has a vaccination card}
\item{dpt3_vx}{binary outcome: vaccinated for DPT3}
\item{mcv_vx}{binary outcome: vaccinated for MCV1}
...
}
}
\usage{
fauxdata
}
\description{
An example dataset containing three survey samples from the imaginary
country of Harmonia
}
\keyword{datasets}
